Add Checkbox component to UI package and enable styled-components support
The Checkbox component has been implemented in the UI package. This involved updates to the UI package's exports in package.json and the creation of new files for the Checkbox, RequiredLabel, and Check components. The styled-components library was also added as a dependency and configured for use in the Next.js web app.