Add Input component to UI package
This commit adds a new Input.tsx file to the UI package, introducing an Input component that can handle customized styling and validation rules. It also updates the package.json file to expose this new component. Minor changes were made to the RequiredLabel component to support the structure of the new Input component.