Add RangeSlider component
A new RangeSlider component is introduced in the UI package. This component allows setting and visualizing a range of numerical values. The change includes the necessary updates to the package.json, navigation menu addition for web app, and corresponding properties updates in codeStrings.ts. A new page to demonstrate the component has also been created.