Add Breadcrumbs component to UI package
The new Breadcrumbs component has been added to the UI package, and is now imported in the web application where it's used to display breadcrumb navigation. This component makes use of the usePathname hook from Next.js's navigation package. The inclusion of this component enhances the navigability of the applications.