Refactor and enhance RangeSlider styling
The styling for the RangeSlider component in the UI package has been improved. Padding has been nested inside the option tag for better clarity. More specific styles, including those related to appearances and dimensions, have been added for better slider visualization and handling in the Webkit and Mozilla browsers. The label attribute from the option tag in the GenerateTickMarks function was also removed as it was unnecessary.