Implement responsive design on Navigation component
The Navigation component in the UI package has been updated to be more responsive. Media queries have been added to adjust layout and visibility for smaller device screen sizes. This includes changes in element padding, alignment, and background color, as well as the addition of abbreviated names for links for improved mobile experience.