Update README to reflect shift to component library
The changes in the README reflect the change in the project's purpose and content. The project is now a collection of (mostly) stand-alone components for Next.js using React, Typescript, and styled components. Instructions have also been updated to use Yarn instead of Turbo commands, and irrelevant sections have been removed.