Add onClose callback and css style support to Notifications
The commit adds an onClose callback function and a css style property to the Notifications component. With these additions, a function can be executed when the notification closes, and additional CSS styles can be applied to the notification. An update to the documentation of the component to reflect these changes is also included in the commit.